moved internal/semver to pkg/semver - renamed semver.detection config to semver.match - removed internal/senver readme
