+++ .semverbot.toml
@@ -10,7 +10,7 @@ prefix = "v"
 [semver]
 mode = "auto"
 
-[semver.detection]
+[semver.match]
 patch = ["fix/", "[fix]"]
 minor = ["feature/", "[feature]"]
 major = ["release/", "[release]"]
