+++ pkg/cli/constants.go
@@ -14,7 +14,7 @@ prefix = "v"
 [semver]
 mode = "auto"
 
-[semver.detection]
+[semver.match]
 patch = ["fix/", "[fix]"]
 minor = ["feature/", "[feature]"]
 major = ["release/", "[release]"]
