+++ pkg/core/predict.go
@@ -1,21 +1,21 @@
 package core
 
 import (
-	"github.com/restechnica/semverbot/internal/semver"
 	"github.com/restechnica/semverbot/pkg/api"
+	"github.com/restechnica/semverbot/pkg/semver"
 )
 
 type PredictVersionOptions struct {
-	DefaultVersion  string
-	SemverDetection map[string][]string
-	SemverMode      string
+	DefaultVersion string
+	SemverMatchMap map[string][]string
+	SemverMode     string
 }
 
 func PredictVersion(options *PredictVersionOptions) (prediction string, err error) {
 	var versionAPI = api.NewVersionAPI()
 	var version = versionAPI.GetVersionOrDefault(options.DefaultVersion)
 
-	var modeDetector = semver.NewModeDetector(options.SemverDetection)
+	var modeDetector = semver.NewModeDetector(options.SemverMatchMap)
 
 	var semverModeAPI = api.NewSemverModeAPI(modeDetector)
 	var semverMode = semverModeAPI.SelectMode(options.SemverMode)
