+++ pkg/semver/auto.go
