+++ pkg/semver/auto_test.go
