+++ pkg/semver/detection.go
@@ -7,20 +7,20 @@ import (
 
 // ModeDetector detects which mode should be applied.
 type ModeDetector struct {
-	ModeDetectionMap map[string][]string
+	SemverMatchMap map[string][]string
 }
 
 // NewModeDetector creates a new ModeDetector with a detection configuration.
 // returns the new ModeDetector.
-func NewModeDetector(modeDetectionMap map[string][]string) ModeDetector {
-	return ModeDetector{ModeDetectionMap: modeDetectionMap}
+func NewModeDetector(semverMatchMap map[string][]string) ModeDetector {
+	return ModeDetector{SemverMatchMap: semverMatchMap}
 }
 
-// DetectMode detects which semver level mode (patch, minor, major) based
-// on a string and the ModeDetectionMap.
+// DetectMode detects the semver level mode (patch, minor, major) based
+// on a string and the SemverMatchMap.
 // Returns the detected mode or an error if no mode was detected.
 func (detector ModeDetector) DetectMode(target string) (detected Mode, err error) {
-	for mode, substrings := range detector.ModeDetectionMap {
+	for mode, substrings := range detector.SemverMatchMap {
 		for _, substring := range substrings {
 			if strings.Contains(target, substring) {
 				switch mode {
