+++ pkg/semver/gitcommit_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/restechnica/semverbot/internal/mocks"
 )
 
-var modeDetectionMap = map[string][]string{
+var semverMatchMap = map[string][]string{
 	Patch: {"[fix]", "fix/"},
 	Minor: {"[feature]", "feature/"},
 	Major: {"[release]", "release/"},
@@ -47,7 +47,7 @@ func TestGitCommitMode_GitCommitConstant(t *testing.T) {
 //		t.Run(test.Name, func(t *testing.T) {
 //			var want = test.Want
 //
-//			var gitCommitMode = NewGitCommitMode(NewModeDetector(modeDetectionMap))
+//			var gitCommitMode = NewGitCommitMode(NewModeDetector(semverMatchMap))
 //			var got, err = gitCommitMode.ModeDetector.(test.Message)
 //
 //			assert.NoError(t, err)
@@ -68,7 +68,7 @@ func TestGitCommitMode_GitCommitConstant(t *testing.T) {
 //		t.Run(test.Name, func(t *testing.T) {
 //			var want = fmt.Sprintf(`could not match a mode to the commit message "%s"`, test.Message)
 //
-//			var gitCommitMode = NewGitCommitMode(modeDetectionMap)
+//			var gitCommitMode = NewGitCommitMode(semverMatchMap)
 //			var _, err = gitCommitMode.GetMatchedMode(test.Message)
 //
 //			assert.Error(t, err)
@@ -101,7 +101,7 @@ func TestGitCommitMode_Increment(t *testing.T) {
 			var cmder = mocks.NewMockCommander()
 			cmder.On("Output", mock.Anything, mock.Anything).Return(test.Message, nil)
 
-			var gitCommitMode = NewGitCommitMode(NewModeDetector(modeDetectionMap))
+			var gitCommitMode = NewGitCommitMode(NewModeDetector(semverMatchMap))
 			gitCommitMode.Commander = cmder
 			var got, err = gitCommitMode.Increment(test.Version)
 
@@ -130,7 +130,7 @@ func TestGitCommitMode_Increment(t *testing.T) {
 			var cmder = mocks.NewMockCommander()
 			cmder.On("Output", mock.Anything, mock.Anything).Return(test.Message, test.GitError)
 
-			var gitCommitMode = NewGitCommitMode(NewModeDetector(modeDetectionMap))
+			var gitCommitMode = NewGitCommitMode(NewModeDetector(semverMatchMap))
 			gitCommitMode.Commander = cmder
 			var _, err = gitCommitMode.Increment(test.Version)
