+++ pkg/semver/mode.go
