+++ pkg/semver/trim.go
